# Conflicts:
#	Starkov.ScheduledReports/Starkov.ScheduledReports.Server/ModuleInitializer.cs
#	Starkov.ScheduledReports/Starkov.ScheduledReports.Server/ModuleServerFunctions.cs
#	Starkov.ScheduledReports/Starkov.ScheduledReports.Server/RelativeDate/RelativeDateServerFunctions.cs
#	Starkov.ScheduledReports/Starkov.ScheduledReports.Shared/Module.mtd
#	Starkov.ScheduledReports/Starkov.ScheduledReports.Shared/Module.resx
#	Starkov.ScheduledReports/Starkov.ScheduledReports.Shared/Module.ru.resx
#	Starkov.ScheduledReports/Starkov.ScheduledReports.Shared/RelativeDate/RelativeDate.mtd

+++ Starkov.ScheduledReports/Starkov.ScheduledReports.Server/ScheduleSetting/ScheduleSettingServerFunctions.cs
@@ -23,7 +23,7 @@ namespace Starkov.ScheduledReports.Server
     /// Получить запись настройки расписания для отчета по ИД.
     /// </summary>
     [Public, Remote(IsPure = true)]
-    public static Starkov.ScheduledReports.IScheduleSetting GetScheduleSetting(int? id)
+    public static Starkov.ScheduledReports.IScheduleSetting GetScheduleSetting(long? id)
     {
       return ScheduleSettings.GetAll(s => s.Id == id).FirstOrDefault(s => s.Status != ScheduledReports.ScheduleSetting.Status.Closed);
     }
